Moves tests for folders in another file
