+++ backend/tracim_backend/tests/functional/test_contents.py
@@ -863,68 +863,6 @@ class TestFolder(FunctionalTest):
         assert "code" in res.json_body
         assert res.json_body["code"] == ErrorCode.GENERIC_SCHEMA_VALIDATION_ERROR
 
-    def test_api__move_folder_into_itself__err_400(self) -> None:
-        self._setup_basics()
-        move_url = "/api/v2/workspaces/{}/contents/{}/move".format(
-            self.workspace.workspace_id, self.folder.content_id
-        )
-        body = {
-            "new_parent_id": self.folder.content_id,
-            "new_workspace_id": self.workspace.workspace_id,
-        }
-        response = self.testapp.put_json(move_url, params=body, status=400)
-        assert response.json_body["code"] == ErrorCode.CONFLICTING_MOVE_IN_ITSELF
-
-    def test_api__move_folder_in_a_direct_child__err_400(self) -> None:
-        self._setup_basics()
-        child_folder = self.content_api.create(
-            label="test_folder",
-            content_type_slug=content_type_list.Folder.slug,
-            workspace=self.workspace,
-            parent=self.folder,
-            do_save=True,
-            do_notify=False,
-        )
-        transaction.commit()
-        move_url = "/api/v2/workspaces/{}/contents/{}/move".format(
-            self.workspace.workspace_id, self.folder.content_id
-        )
-        body = {
-            "new_parent_id": child_folder.content_id,
-            "new_workspace_id": self.workspace.workspace_id,
-        }
-        response = self.testapp.put_json(move_url, params=body, status=400)
-        assert response.json_body["code"] == ErrorCode.CONFLICTING_MOVE_IN_CHILD
-
-    def test_api__move_folder_in_a_sub_child__err_400(self) -> None:
-        self._setup_basics()
-        child_folder = self.content_api.create(
-            label="child_test_folder",
-            content_type_slug=content_type_list.Folder.slug,
-            workspace=self.workspace,
-            parent=self.folder,
-            do_save=True,
-            do_notify=False,
-        )
-        sub_child_folder = self.content_api.create(
-            label="sub_child_test_folder",
-            content_type_slug=content_type_list.Folder.slug,
-            workspace=self.workspace,
-            parent=child_folder,
-            do_save=True,
-            do_notify=False,
-        )
-        transaction.commit()
-        move_url = "/api/v2/workspaces/{}/contents/{}/move".format(
-            self.workspace.workspace_id, self.folder.content_id
-        )
-        body = {
-            "new_parent_id": sub_child_folder.content_id,
-            "new_workspace_id": self.workspace.workspace_id,
-        }
-        response = self.testapp.put_json(move_url, params=body, status=400)
-        assert response.json_body["code"] == ErrorCode.CONFLICTING_MOVE_IN_CHILD
-
 
 class TestHtmlDocuments(FunctionalTest):
     """
